+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+   "C_Cpp.default.compilerPath": "C:/msys64/mingw64/bin/g++.exe"
+}
