+++ include/core/Flight.h
@@ -1,5 +1,4 @@
 #pragma once
-
 #include <string>
 
 struct Flight {
