+++ include/services/BookingService.h
@@ -5,6 +5,7 @@
 #include <vector>
 
 class BookingService {
+   Flight flight;
    FlightStorage storage;
 public:
    void bookFlight();
@@ -12,7 +13,7 @@ public:
    std::vector<std::string> getValidFlights(std::string depCity, std::string arrCity);
    std::vector<std::string> getAllFlights();
 
-   std::optional<Flight> getFlight(std::string id);
+   std::optional<Flight> getFlight(std::string ID);
 
    friend bool isValidCity(std::string city, BookingService& booker);
 };
