+++ src/storage/FlightStorage.cpp
@@ -3,6 +3,7 @@
 #include <vector>
 #include <string>
 #include <fstream>
+#include <optional>
 
 FlightStorage::FlightStorage(){}
 FlightStorage::FlightStorage(const std::string& BASE_PATH){
